Create ReNamer.py
The script prompts the user for the location of a folder containing files they want to rename and the location of a folder to save the new files to. It then prompts the user for the main name that the files will share and appends the date and time the file was created as a unique identifier to the main name. It then renames and saves the files to the specified folder while also adding error checking to make sure the inputs and file operations are valid.